Optimized function for static assets
